refactor: group core module files into features
