+++ projects/ngx-meta/src/core/index.ts
@@ -1,26 +1,31 @@
-// Standalone & module entry points
-export * from './src/ngx-meta-core.module'
-export * from './src/ngx-meta-metadata-loader.module'
-export * from './src/provide-ngx-meta-core'
-export * from './src/provide-ngx-meta-metadata-loader'
-export * from './src/with-ngx-meta-defaults'
-// Services & utils
-export * from './src/make-composed-key-val-meta-definition'
-export * from './src/make-key-val-meta-definition'
-export * from './src/make-metadata-manager-provider-from-setter-factory'
-export * from './src/ngx-meta-meta.service'
-export * from './src/ngx-meta.service'
-// Types
-export * from './src/global-metadata'
-export * from './src/global-metadata-image'
+// Feature providers
+export * from './src/providers/ngx-meta-core.module'
+export * from './src/providers/provide-ngx-meta-core'
+// Defaults
+export * from './src/defaults/with-ngx-meta-defaults'
+// Globals
+export * from './src/globals/global-metadata'
+export * from './src/globals/global-metadata-image'
+// Head elements
+export * from './src/head-elements/head-element-upsert-or-remove'
+// Loader
+export * from './src/loader/ngx-meta-metadata-loader.module'
+export * from './src/loader/provide-ngx-meta-metadata-loader'
+// Managers
+export * from './src/managers/make-metadata-manager-provider-from-setter-factory'
+export * from './src/managers/ngx-meta-metadata-manager'
+// Messaging (internal)
+export * from './src/messaging/format-dev-message'
+export * from './src/messaging/maybe-non-http-url-dev-message'
+export * from './src/messaging/maybe-too-long-dev-message'
+// Meta elements
+export * from './src/meta-elements/make-composed-key-val-meta-definition'
+export * from './src/meta-elements/make-key-val-meta-definition'
+export * from './src/meta-elements/ngx-meta-meta-content'
+export * from './src/meta-elements/ngx-meta-meta-definition'
+export * from './src/meta-elements/ngx-meta-meta.service'
+// Routing (internal)
+export * from './src/routing/route-metadata-strategy'
+// General
 export * from './src/metadata-values'
-export * from './src/ngx-meta-meta-content'
-export * from './src/ngx-meta-meta-definition'
-// Tokens
-export * from './src/ngx-meta-metadata-manager'
-// Internal
-export * from './src/head-element-upsert-or-remove'
-export * from './src/route-metadata-strategy'
-export * from './src/format-dev-message'
-export * from './src/maybe-non-http-url-dev-message'
-export * from './src/maybe-too-long-dev-message'
+export * from './src/ngx-meta.service'
