+++ projects/ngx-meta/src/core/src/globals/global-metadata-image.ts
