+++ projects/ngx-meta/src/core/src/globals/global-metadata.ts
