+++ projects/ngx-meta/src/core/src/loader/ngx-meta-metadata-loader.module.ts
