+++ projects/ngx-meta/src/core/src/managers/__tests__/make-metadata-manager-spy.ts
