+++ projects/ngx-meta/src/core/src/managers/make-metadata-manager-provider-from-setter-factory.ts
