+++ projects/ngx-meta/src/core/src/managers/metadata-registry.spec.ts
