+++ projects/ngx-meta/src/core/src/managers/metadata-registry.ts
