+++ projects/ngx-meta/src/core/src/managers/ngx-meta-metadata-manager.ts
