+++ projects/ngx-meta/src/core/src/messaging/maybe-non-http-url-dev-message.ts
