+++ projects/ngx-meta/src/core/src/meta-elements/ngx-meta-meta-content.ts
