+++ projects/ngx-meta/src/core/src/providers/ngx-meta-core.module.spec.ts
@@ -1,9 +1,9 @@
 import { TestBed } from '@angular/core/testing'
 import { NgxMetaCoreModule } from './ngx-meta-core.module'
-import { injectDefaults } from './defaults'
+import { injectDefaults } from '../defaults/defaults'
 import { ModuleWithProviders } from '@angular/core'
-import { GlobalMetadata } from './global-metadata'
-import { withNgxMetaDefaults } from './with-ngx-meta-defaults'
+import { GlobalMetadata } from '../globals/global-metadata'
+import { withNgxMetaDefaults } from '../defaults/with-ngx-meta-defaults'
 
 describe('Core module', () => {
   const defaults: GlobalMetadata = { title: 'Hello World!' }
