+++ projects/ngx-meta/src/core/src/providers/ngx-meta-core.module.ts
@@ -1,6 +1,6 @@
 import { ModuleWithProviders, NgModule } from '@angular/core'
-import { MetadataValues } from './metadata-values'
-import { withNgxMetaDefaults } from './with-ngx-meta-defaults'
+import { MetadataValues } from '../metadata-values'
+import { withNgxMetaDefaults } from '../defaults/with-ngx-meta-defaults'
 import {
   CoreFeature,
   CoreFeatureKind,
