+++ projects/ngx-meta/src/core/src/routing/route-metadata-strategy.ts
@@ -1,5 +1,5 @@
 import { inject, InjectionToken } from '@angular/core'
-import { MetadataValues } from './metadata-values'
+import { MetadataValues } from '../metadata-values'
 
 /**
  * @internal
