+++ projects/ngx-meta/src/core/src/utils/is-object.ts
